+++ ergo/ChangeLog.md
@@ -8,6 +8,10 @@
 * Fix a minor bug in doc comment leading whitespace removal.
 * Improve error message printing and information.
 * Improve terminal UI rendering behavior.
+* Add support for tree comments.
+  * If a comment token (`#`) is followed by whitespace it is a line comment,
+    otherwise it is a tree comment. A tree comment will comment out the
+    following syntax tree.
 
 ## 1.0.0-rc.0  -- 2021-08-01
 Most of the changes listed below are breaking changes, so there is no separate